Refactored frames in their own modules
